LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/browsers/windows/browsing-context-names/choose-_top-001-expected.txt

CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.

Harness Error (TIMEOUT), message = null

TIMEOUT Should choose current browsing context for "_top" if current is top Test timed out
